refactor!: move role parsers to guild file
