+++ src/disnake/ext/components/impl/parser/__init__.py
@@ -12,6 +12,5 @@ from disnake.ext.components.impl.parser.emoji import *
 from disnake.ext.components.impl.parser.enum import *
 from disnake.ext.components.impl.parser.guild import *
 from disnake.ext.components.impl.parser.message import *
-from disnake.ext.components.impl.parser.role import *
 from disnake.ext.components.impl.parser.snowflake import *
 from disnake.ext.components.impl.parser.user import *
